requirements.txt
apache-beam[gcp]==2.60.0
google-cloud-bigquery==3.30.0
google-cloud-pubsub==2.29.0
faker==37.1.0

